minor upd to pass forge fmt
minor upd to pass forge fmt

+++ src/interfaces/IWETH9_066.sol
@@ -3,6 +3,6 @@ pragma solidity ^0.8.20;
 
 interface IWETH9_066 {
     function deposit() external payable;
-    function withdraw(uint wad) external;
-    function approve(address guy, uint wad) external returns (bool);
+    function withdraw(uint256 wad) external;
+    function approve(address guy, uint256 wad) external returns (bool);
 }
